fix: move vue to the content-script page
